+++ src/SCSS/_base.scss
@@ -6,3 +6,6 @@ body {
   margin: 0;
   overflow: hidden;
 }
+* {
+  outline: none !important;
+}
