+++ src/SCSS/_variables.scss
@@ -1,10 +1,16 @@
 $color-primary: rgb(59, 189, 176);
-$color-accent: rgb(192, 189, 189);
-$color-row-hover: rgba(192, 189, 189, 0.5);
+$color-accent: rgb(144, 221, 240);
+$color-row-hover: rgb(232, 236, 236);
 $color-black: black;
 $color-white: white;
 
 $color-text-primary: rgb(129, 127, 127);
+$color-text-heading: rgb(41, 119, 110);
 
 $complete-width: 100%;
-$viewport-2: 2vh;
+$viewport-basic: 2vh;
+
+$profile-picture-background: rgba(192, 189, 189, 0.5);
+
+$width-image-container: 20%;
+$width-container: 80%;
