--- src/app/_helpers/db.json
@@ -1,14 +1,0 @@
-{
-  "candidates": [
-    {
-      "id": 1,
-      "candidateId": 1,
-      "name": "Himan",
-      "email": "himan@abc.xyz",
-      "phone": 1234567890,
-      "experience": 4.5,
-      "photo": "assets/profile_pic_default.jpg",
-      "github": "https://www.github.com/him123"
-    }
-  ]
-}
