+++ src/app/candidate/all-candidates/all-candidates.component.html
@@ -35,7 +35,9 @@
 
     <ng-container matColumnDef="phone">
       <th mat-header-cell *matHeaderCellDef>Phone</th>
-      <td mat-cell *matCellDef="let candidate">+{{ candidate.phone }}</td>
+      <td mat-cell *matCellDef="let candidate">
+        +{{ candidate.code }}-{{ candidate.phone }}
+      </td>
     </ng-container>
 
     <ng-container matColumnDef="experience">
