+++ src/app/candidate/all-candidates/all-candidates.component.scss
@@ -11,11 +11,12 @@ main {
   & > table {
     width: $complete-width;
     height: 20%;
-    & > tr {
+    tr {
       color: $color-text-primary;
 
       &:hover {
         background-color: $color-row-hover;
+        cursor: pointer;
       }
     }
 
@@ -24,13 +25,13 @@ main {
     }
 
     td.mat-cell {
-      font-size: $viewport-2;
+      font-size: $viewport-basic;
     }
 
     .mat-header-cell {
       color: $color-black;
-      background-color: $color-accent;
-      font-size: $viewport-2;
+      background-color: $color-row-hover;
+      font-size: $viewport-basic;
     }
     & .nameColumn {
       color: $color-primary;
@@ -41,7 +42,7 @@ main {
     width: 88%;
   }
   & > .sidebar {
-    padding: 0 $viewport-2;
+    padding: 0 $viewport-basic;
 
     & > button {
       background-color: $color-primary;
