+++ src/app/candidate/new-candidate/new-candidate.component.scss
@@ -1,67 +1,77 @@
+@import '../../../styles.scss';
+
 .gobackContainer {
   padding: 0 5%;
 
-  & > span:first-child {
+  & > span:first-child,
+  fa-icon {
+    cursor: pointer;
+    font-size: 1.2rem;
+    color: $color-text-primary;
+  }
+  label {
     cursor: pointer;
   }
 }
 
 main {
   padding: 0 5%;
-  display: flex;
-  justify-content: center;
-
-  header {
-    font-size: 3vh;
-    padding-bottom: 2vh;
+  @include centralizeContent;
+  & header {
+    font-size: 1.5rem;
+    margin-bottom: $viewport-basic;
     text-align: center;
+    color: $color-text-primary;
   }
 }
 
 .picUpload {
-  display: block;
-  height: 3.5vh;
-  width: 5vh;
-  padding: 5vh;
-  padding-top: 6vh;
-  border-radius: 50%;
-  background-color: rgba(192, 189, 189, 0.5);
+  @include profilePicture;
+  @include centralizeContent;
+  align-items: center;
+  background-color: $profile-picture-background;
+  margin-bottom: 2vh;
+
+  & > label {
+    cursor: pointer;
+  }
 }
 img {
-  height: 15vh;
-  width: 15vh;
-  border-radius: 50%;
+  @include profilePicture;
+  object-fit: cover;
+  cursor: pointer;
 }
 
 .candidate-field-form {
-  min-width: 150px;
-  max-width: 500px;
-  width: 100%;
+  min-width: 20vw;
+  max-width: 30vw;
+  width: $complete-width;
 }
 
 .candidate-field-full-width {
-  width: 100%;
-}
-
-.phoneCodeClass {
-  width: 20%;
-}
-
-.phoneNumberClass {
-  width: 75%;
-  margin-left: 1vh;
+  width: $complete-width;
 }
 
 .row {
   display: flex;
+
+  & .phoneCodeClass {
+    width: 20%;
+  }
+
+  & .phoneNumberClass {
+    width: 75%;
+    margin-left: 1vh;
+  }
 }
 
 button {
-  font-size: 2vh;
-  width: 100%;
+  font-size: $viewport-basic;
+  width: $complete-width;
   height: 5vh;
   cursor: pointer;
-}
-button:disabled {
-  cursor: not-allowed;
+
+  &:disabled {
+    cursor: not-allowed;
+  }
 }
