+++ src/app/candidate/profile/profile.component.scss
@@ -1,30 +1,37 @@
+@import '../../../styles.scss';
+
 main {
   display: flex;
   justify-content: flex-start;
+
+  & .imageContainer {
+    width: $width-image-container;
+    img {
+      @include profilePicture;
+      object-fit: cover;
+    }
+  }
+  & .container {
+    width: $width-container;
+    margin: 0 $viewport-basic $viewport-basic 3vh;
+    & > h1 {
+      margin-block-start: 0;
+    }
+
+    & .gridContainer {
+      display: grid;
+      grid-template-columns: auto auto;
+
+      & span {
+        display: inline-block;
+        padding: 1vh;
+      }
+    }
+  }
 }
-.imageContainer {
-  width: 20%;
-}
-img {
-  height: 15vh;
-  width: 15vh;
-  border-radius: 50%;
-}
-.container {
-  width: 80%;
-  margin: 0 2vh 2vh 3vh;
-}
-.gridContainer {
-  display: grid;
-  grid-template-columns: auto auto;
-}
-h1 {
-  margin-block-start: 0;
-}
-span {
-  display: inline-block;
-  padding: 1vh;
-}
-i {
-  color: black;
+
+span,
+i,
+fa-icon {
+  color: $color-text-primary;
 }
