+++ src/app/candidate/profile/profile.component.ts
@@ -1,9 +1,7 @@
 import { Component, OnInit, Inject } from '@angular/core';
-import { DataService } from 'src/app/_helpers/data.service';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { CandidateElement } from 'src/app/_helpers/candidate-interface';
+import { DataService } from 'src/app/_collaborators/data.service';
+import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { faMobileAlt, faEnvelope, faBriefcase } from '@fortawesome/free-solid-svg-icons';
-import {} from '@fortawesome/fontawesome-svg-core';
 
 @Component({
   selector: 'shortlist-profile',
@@ -11,21 +9,18 @@ import {} from '@fortawesome/fontawesome-svg-core';
   styleUrls: ['./profile.component.scss']
 })
 export class ProfileComponent implements OnInit {
-  dataInput;
+  dataId;
   faMobileAlt = faMobileAlt;
   faEnvelope = faEnvelope;
   faBriefcase = faBriefcase;
-  constructor(
-    private dataService: DataService,
-    private dialogRef: MatDialogRef<CandidateElement>,
-    @Inject(MAT_DIALOG_DATA) data
-  ) {
-    this.dataInput = data.id;
+  constructor(private dataService: DataService, @Inject(MAT_DIALOG_DATA) data) {
+    this.dataId = data.id;
   }
   profileData;
-  id;
   ngOnInit() {
-    this.profileData = this.dataService.getDataById(this.dataInput);
-    document.getElementById('profilePhoto').setAttribute('src', this.profileData.photo);
+    this.dataService.getDataById(this.dataId).subscribe(el => {
+      this.profileData = el;
+      document.getElementById('profilePhoto').setAttribute('src', this.profileData.photo);
+    });
   }
 }
