+++ src/app/footer/footer.component.scss
@@ -1,5 +1,5 @@
+@import '../../styles.scss';
+
 footer {
-  display: flex;
-  justify-content: center;
-  padding: 2rem;
+  @include header-footer-mixin;
 }
