+++ src/styles.scss
@@ -2,3 +2,4 @@
 @import './SCSS/base';
 @import './SCSS/variables';
 @import './SCSS/typography';
+@import './SCSS/mixin';
